Move method to kiam_helper file
