+++ smoke-tests/spec/spec_helper.rb
@@ -104,6 +104,7 @@ require "kubernetes_helper"
 require "erb"
 require "json"
 require "date"
+require "./spec/kiam_helper"
 
 def readable_timestamp
   Time.now.strftime("%Y%m%d%H%M%S")
